+++ src/main/resources/db/migration/V1__init_db.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS CRYPTO_CURRENCY
 (
     currency_name   varchar(255) PRIMARY KEY,
-    currency_value  double
+    currency_value  numeric
 
 );
